Refactoring of the code and documentation
