+++ proteobench/modules/dda_id/module_dda_id.py
@@ -1,0 +1,7 @@
+""" Main interface of the module."""
+
+
+def is_implemented() -> bool:
+    """ Returns whether the module is fully implemented. """
+    return False
+
