+++ proteobench/modules/dda_quant/io_parse_settings/parse_settings_maxquant.toml
@@ -14,12 +14,12 @@ LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_02 = 2
 LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_03 = 2
 
 [run_mapper]
-LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_01 = LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_01
-LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_02 = LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_02
-LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_03 = LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_03
-LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_01 = LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_01
-LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_02 = LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_02
-LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_03 = LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_03
+LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_01 = "LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_01"
+LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_02 = "LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_02"
+LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_03 = "LFQ_Orbitrap_DDA_Condition_A_Sample_Alpha_03"
+LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_01 = "LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_01"
+LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_02 = "LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_02"
+LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_03 = "LFQ_Orbitrap_DDA_Condition_B_Sample_Alpha_03"
 
 [species_dict]
 "YEAST" = "_YEAST"
