+++ proteobench/modules/dia_quant/module_dia_quant.py
@@ -1,0 +1,7 @@
+""" Main interface of the module."""
+
+
+def is_implemented() -> bool:
+    """ Returns whether the module is fully implemented. """
+    return False
+
