+++ test/test_module_dda_quant.py
@@ -1,7 +1,8 @@
 import os
 import unittest
 
-import proteobench.modules.dda_quant as dda_quant
+from proteobench.modules.dda_quant import module_dda_quant
+from proteobench.modules.dda_quant.parse_settings_dda_quant import INPUT_FORMATS
 
 TESTDATA_DIR = os.path.join(os.path.dirname(__file__), 'data')
 TESTDATA_FILES = { "WOMBAT"         : os.path.join(TESTDATA_DIR, 'WOMBAT_stand_pep_quant_mergedproline.csv'),
@@ -9,23 +10,15 @@ TESTDATA_FILES = { "WOMBAT"         : os.path.join(TESTDATA_DIR, 'WOMBAT_stand_p
                    "MSFragger"         : os.path.join(TESTDATA_DIR, 'MSFragger_combined_ion.tsv')
             }
 
-INPUT_FORMATS = {   "MaxQuant" : "MaxQuant",
-                    "AlphaPept" : "AlphaPept",
-                    "MSFragger" : "MSFragger",
-                    "Proline" : "Proline",
-                    "WOMBAT" : "WOMBAT"}
-
 
 def load_file(format_name:str):
         """ Method used to load the input file of a given format."""
         user_input = dict()
         user_input["input_csv"] = TESTDATA_FILES[format_name]
-        user_input["input_format"] = INPUT_FORMATS[format_name]
-        user_input["mbr"] = True
-        df = dda_quant.main(
+        user_input["input_format"] = format_name
+        df = module_dda_quant.main(
             user_input["input_csv"],
-            user_input["input_format"],
-            user_input["mbr"]
+            user_input["input_format"]
         )
         return df
 
@@ -34,18 +27,21 @@ class TestOutputFileReading(unittest.TestCase):
     def test_MaxQuant_file(self):
         """ Test whether MaxQuant input is parsed correctly."""
         test_dataset_name = "MaxQuant"
+        self.assertTrue(test_dataset_name in INPUT_FORMATS)
         df = load_file(test_dataset_name)
         self.assertFalse(df.empty)
 
     def test_Wombat_file(self):
         """ Test whether WOMBAT input is parsed correctly."""
         test_dataset_name = "WOMBAT"
+        self.assertTrue(test_dataset_name in INPUT_FORMATS)
         df = load_file(test_dataset_name)
         self.assertFalse(df.empty)
 
     def test_MSFragger_file(self):
         """ Test whether MSFragger input is parsed correctly."""
         test_dataset_name = "MSFragger"
+        self.assertTrue(test_dataset_name in INPUT_FORMATS)
         df = load_file(test_dataset_name)
         self.assertFalse(df.empty)
